16th Video Done - installed django_filters
Add filtering support for Product API

Integrates django-filters to enable query-based filtering in the Product API.
Adds a ProductFilter class to define filterable fields for
ame and price.
Configures the ProductListCreateAPIView to use the filterset and updates settings
to include django_filters in installed apps and as a default filter backend.